pyproject.toml
[project]
name = "adk-web-chat-ui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "google-adk>=1.13.0",
    "litellm>=1.76.3",
    "loguru>=0.7.3",
    "pydantic-settings>=2.10.1",
    "pyjwt[crypto]>=2.10.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "ujson>=5.11.0",
]

[dependency-groups]
dev = [
    "ruff>=0.12.12",
]

[[tool.uv.index]]
url = "http://nexus.wangankeji.com/repository/pypi.tsinghua-proxy/simple"
default = true

[tool.ruff]
target-version = "py313"
line-length = 100
indent-width = 4

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
    "W191", # indentation contains tabs
]

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true
